edits for python 2+3 dual compatibility: (1) add from __future__ import division, print_function, absolute_import to every python file, so that py2 and py3 have mostly the same behavior & bugs; (2) print --> print(); (3) xrange --> range [it has no performance impact for small ranges]; (4) repair floor integer divisions to keep the same behavior (at least the ones I can find...); (5) implicit relative imports --> explicit; (6) open(...'wb') --> open(...'w') in install.py for py3 compatibility.

+++ examples/Silicon_Grating_Coupler/gc_opt_constrained.py
@@ -43,6 +43,7 @@ in the command line which will run the optimization using 16 cores on the curren
 machine.
 
 """
+from __future__ import division, print_function, absolute_import
 # We need to import a lot of things from emopt
 import emopt
 from emopt.misc import NOT_PARALLEL
@@ -132,7 +133,7 @@ class SiliconGratingAM(AdjointMethodPNF):
         This generally works pretty well, as well.
         """
         coeffs = params
-        N_coeffs = (len(coeffs) - 3) / 4
+        N_coeffs = int((len(coeffs) - 3) / 4)
 
         h_etch = params[-3]
         x0 = self.w_in + params[-2]
